Refactor common build-tools classes
