+++ src/main/java/org/jfrog/hudson/pipeline/common/types/builds/NpmBuild.java
@@ -4,7 +4,7 @@ import com.google.common.collect.Maps;
 import org.jenkinsci.plugins.scriptsecurity.sandbox.whitelists.Whitelisted;
 import org.jfrog.hudson.pipeline.common.types.buildInfo.BuildInfo;
 import org.jfrog.hudson.pipeline.common.types.deployers.NpmGoDeployer;
-import org.jfrog.hudson.pipeline.common.types.resolvers.NpmGoResolver;
+import org.jfrog.hudson.pipeline.common.types.resolvers.CommonResolver;
 
 import java.util.Arrays;
 import java.util.List;
@@ -24,7 +24,7 @@ public class NpmBuild extends PackageManagerBuild {
 
     public NpmBuild() {
         deployer = new NpmGoDeployer();
-        resolver = new NpmGoResolver();
+        resolver = new CommonResolver();
     }
 
     @Whitelisted
