+++ src/main/java/org/jfrog/hudson/pipeline/declarative/steps/go/GoResolverStep.java
@@ -2,18 +2,18 @@ package org.jfrog.hudson.pipeline.declarative.steps.go;
 
 import hudson.Extension;
 import org.jenkinsci.plugins.workflow.steps.AbstractStepDescriptorImpl;
-import org.jfrog.hudson.pipeline.declarative.steps.common.NpmGoDeployerResolver;
-import org.jfrog.hudson.pipeline.common.types.resolvers.NpmGoResolver;
+import org.jfrog.hudson.pipeline.declarative.steps.common.DeployerResolverBase;
+import org.jfrog.hudson.pipeline.common.types.resolvers.CommonResolver;
 import org.kohsuke.stapler.DataBoundConstructor;
 
-public class GoResolverStep extends NpmGoDeployerResolver {
+public class GoResolverStep extends DeployerResolverBase {
 
     static final String STEP_NAME = "rtGoResolver";
 
     @DataBoundConstructor
     public GoResolverStep(String id, String serverId, String repo) {
         super(STEP_NAME, id, serverId);
-        NpmGoResolver goResolver = new NpmGoResolver();
+        CommonResolver goResolver = new CommonResolver();
         goResolver.setRepo(repo);
         buildDataFile.putPOJO(goResolver);
     }
