+++ src/main/java/org/jfrog/hudson/pipeline/declarative/steps/go/GoRunStep.java
@@ -10,7 +10,7 @@ import org.jfrog.hudson.pipeline.common.types.ArtifactoryServer;
 import org.jfrog.hudson.pipeline.ArtifactorySynchronousNonBlockingStepExecution;
 import org.jfrog.hudson.pipeline.common.types.buildInfo.BuildInfo;
 import org.jfrog.hudson.pipeline.common.types.builds.GoBuild;
-import org.jfrog.hudson.pipeline.common.types.resolvers.NpmGoResolver;
+import org.jfrog.hudson.pipeline.common.types.resolvers.CommonResolver;
 import org.jfrog.hudson.pipeline.declarative.BuildDataFile;
 import org.jfrog.hudson.pipeline.declarative.utils.DeclarativePipelineUtils;
 import org.jfrog.hudson.util.BuildUniqueIdentifierHelper;
@@ -99,7 +99,7 @@ public class GoRunStep extends AbstractStepImpl {
             if (buildDataFile == null) {
                 throw new IOException("Resolver " + step.resolverId + " doesn't exist!");
             }
-            NpmGoResolver resolver = SerializationUtils.createMapper().treeToValue(buildDataFile.get(GoResolverStep.STEP_NAME), NpmGoResolver.class);
+            CommonResolver resolver = SerializationUtils.createMapper().treeToValue(buildDataFile.get(GoResolverStep.STEP_NAME), CommonResolver.class);
             resolver.setServer(getArtifactoryServer(buildNumber, buildDataFile));
             step.goBuild.setResolver(resolver);
             addProperties(buildDataFile);
