+++ src/main/java/org/jfrog/hudson/pipeline/declarative/steps/npm/NpmDeployerStep.java
@@ -3,7 +3,7 @@ package org.jfrog.hudson.pipeline.declarative.steps.npm;
 import hudson.Extension;
 import org.jenkinsci.plugins.workflow.steps.AbstractStepDescriptorImpl;
 import org.jfrog.hudson.pipeline.common.types.deployers.NpmGoDeployer;
-import org.jfrog.hudson.pipeline.declarative.steps.common.NpmGoDeployerResolver;
+import org.jfrog.hudson.pipeline.declarative.steps.common.DeployerResolverBase;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
 
@@ -12,7 +12,7 @@ import java.util.List;
 /**
  * @author yahavi
  */
-public class NpmDeployerStep extends NpmGoDeployerResolver {
+public class NpmDeployerStep extends DeployerResolverBase {
 
     static final String STEP_NAME = "rtNpmDeployer";
     private NpmGoDeployer npmDeployer;
